# fix: improve uint64 perf