--- a/packages/protons-runtime/package.json
+++ b/packages/protons-runtime/package.json
@@ -126,6 +126,7 @@
     "release": "aegir release"
   },
   "dependencies": {
+    "uint8-varint": "^2.0.2",
     "uint8arraylist": "^2.4.3",
     "uint8arrays": "^4.0.6"
   },