@achingbrain 2024-01-31
Running `/packages/protons-benchmark/src/numbers/index.ts` shows a nice speed up for `uint64number` performance:

Before:

```
-- read --
uint64 (BigInt) x 73,711,416 ops/sec ±0.57% (88 runs sampled)
uint64number x 75,126,986 ops/sec ±0.26% (96 runs sampled)
uint64string x 31,420,170 ops/sec ±0.43% (94 runs sampled)
Fastest is uint64number
-- write --
uint64 (BigInt) x 41,097,033 ops/sec ±0.61% (94 runs sampled)
uint64number x 89,523,652 ops/sec ±0.60% (97 runs sampled)
uint64string x 18,610,059 ops/sec ±0.19% (98 runs sampled)
Fastest is uint64number
```

After:

```
-- read --
uint64 (BigInt) x 73,310,378 ops/sec ±0.60% (91 runs sampled)
uint64number x 134,356,515 ops/sec ±0.48% (95 runs sampled)
uint64string x 31,575,496 ops/sec ±0.53% (90 runs sampled)
Fastest is uint64number
-- write --
uint64 (BigInt) x 41,444,957 ops/sec ±0.47% (95 runs sampled)
uint64number x 114,640,310 ops/sec ±0.81% (94 runs sampled)
uint64string x 18,531,535 ops/sec ±0.37% (97 runs sampled)
Fastest is uint64number
```